dry-validation 'Optional keys and values' updated
You state that you 'make a clear distinction between specifying an optional `key` and an optional `value`' but the description below is actually misleading - the age key in the example is optional, but if it is present, the value MUST be filled. It can NOT be nil as you say.